file: Keep lister helpers outside of list_directory()
The linux_dirent64 and dirent_type() helper, that converts linux dirent
into seastar direct, will be reused later
