+++ include/seastar/core/file-types.hh
@@ -78,6 +78,24 @@ enum class directory_entry_type {
     socket,
 };
 
+namespace internal::linux_abi {
+
+// From getdents(2):
+// check for 64-bit inode number
+static_assert(sizeof(ino_t) == 8, "large file support not enabled");
+static_assert(sizeof(off_t) == 8, "large file support not enabled");
+
+// From getdents(2):
+struct linux_dirent64 {
+    ino64_t        d_ino;    /* 64-bit inode number */
+    off64_t        d_off;    /* 64-bit offset to next structure */
+    unsigned short d_reclen; /* Size of this dirent */
+    unsigned char  d_type;   /* File type */
+    char           d_name[]; /* Filename (null-terminated) */
+};
+
+} // internal::linux_abi namespace
+
 /// Enumeration describing the type of a particular filesystem
 enum class fs_type {
     other,
